Create and wire in the concept of a Log Consumer
The Log Controller ensures the /logs parameters are valid and parsed
to its knowledge.

- filename must be present
- lines (if present) must be parsed as an integer

Then LogConsumer.Options.build will apply its domain specific
validation of the options

- lines must be an integer >= 1

That isn't a controller validation because it isn't directly a
requirement of the API contract itself: it's possible that Log
Consumer will allow negative values in the future to, say, get lines
from the start of a file.

Once the options are validated and built into a struct of options, the
Log Controller calls the Log Consumer with those options.

The Log Consumer will spin off work via the Task.Supervisor to collect,
filter, and return log lines for given files.

As of this commit the /logs API responds but only returns a hardcoded
set of four lines for any request.

But now the request parameters are parsed and converted to an Options
struct and we are now ready to actually either start implementing tests
or start actually talking to the filesystem.

Example API calls
-----------------

```
$ curl localhost:4000/api/logs # invalid: no filename
$ curl localhost:4000/api/logs\?filename=system.log
$ curl localhost:4000/api/logs\?filename=system.log\&lines=50\&filter=testing
```