Expand LogConsumer to actually read from files
I also added a configuration setting to allow setting the log file
path from ENV or use a default. In development the default log
directory will be ./sample_logs

NOTE: We need to handle security restrictions. Users should not be
able to use a filename like "../../../etc/passwd" to get to files
outside of the configured directory.

NOTE: This approach is also sadly not performant for large log files.
It seems that File.stream is indeed not a panacea for reading from
large files.