Read from the END of log files for performance
My initial naive approach of using File.stream!/1 worked great…on
small files. Unfortunately the name "stream" isn't a magic word to
avoid performance considerations. All the typical high level file
functions want to read a file from the start of the file. If we were
going to consume the entire file and wanted to do so efficiently
(in terms of memory) then File.stream!/1 would let us do that.

But Blammo does not want to read from the start of the file. It wants
to tail/grep for lines at the end of the file because we're (very
reasonably) assuming the log files we're reading are being being
appended to and so the newest lines are at the end.

What's the approach here? Well it's an approach that isn't optimal yet
that's for sure. But it is fast (enough) and has stable memory usage
even under load.

The approach
------------

- have a file
- find the end of the file in bytes (the file size)
- have a first guess at a chunk size
- start reading from (the end of the file) - (the chunk size)
- read until the end of the file
- split the read bytes into lines
- take the requested limit from those lines

If we got the requested limit of lines back: great we're done!

If we had less lines than we wanted: double the chunk size and try
again.

(NOTE: this is an optimization waiting to happen. We are reading the
same bytes repeatedly.)

(NOTE: this is also unguarded as of this commit: we should probably
stop the search at some number of iterations or we could end up reading
the entire file from finish to start.)

The order of operations
-----------------------

There are two ways we could go about our log reading and filtering

1. Find the limit of lines requested first, then filter the results.

   This is fast and trivial.

2. Filter the results until we reach the limit of lines requested.

   This is less fast and has some ramifications to watch out for. e.g.
   what if the pattern you are filtering for happens in the first line
   of the file?

I have written code to support both options.

For option 1 the LogsController should call

```elixir
Blammo.LogConsumer.consume_lines_first/1
```

For option 2 the LogsController should call

```elixir
Blammo.LogConsumer.consume_filter_first/1
```

Performance
-----------

Both approaches are reasonably performant and memory stable.

Option 1 example request timing against a 1GB log file

```
[debug] Processing with BlammoWeb.LogsController.loglines/2
  Parameters: %{"filename" => "sample.1GB.log", "filter" => "xyzzy", "limit" => "1000"}
  Pipelines: [:api]
[info] Sent 200 in 906µs
```

Options 2 example request timing against a 1GB log file

```
[debug] Processing with BlammoWeb.LogsController.loglines/2
  Parameters: %{"filename" => "sample.1GB.log", "filter" => "xyzzy", "limit" => "1000"}
  Pipelines: [:api]
[info] Sent 200 in 7ms
```

hey benchmarking
----------------

```
hey http://localhost:4000/api/logs\?filename\=sample.1GB.log\&lines\=1000\&filter\=xyzzy
```

Option 1

Summary:
  Total:	0.1003 secs
  Slowest:	0.0373 secs
  Fastest:	0.0114 secs
  Average:	0.0231 secs
  Requests/sec:	1993.9434

Option 2

Summary:
  Total:	0.2709 secs
  Slowest:	0.0900 secs
  Fastest:	0.0298 secs
  Average:	0.0627 secs
  Requests/sec:	738.4097