Add APIs for choosing tail or filter first
Blammo allows filtering the log until it finds enough lines to reach
the given limit OR getting the limit of lines and then filtering.

The first approach is slower and has no guarantee that it won't have
to read the entire file: but it is arguably useful.

The second approach is faster and will only read the entire file if
the user requests that many lines.