Improve performance for the filter-first path
Now instead of reading from the end of the file with an increasingly
large chunk we step backwards through the file in chunks until we
either have enough lines or we reach the beginning of the file.

The chunk size will increase until it hits a maximum size and from
that point will remain the same size until it reaches the beginning
of the file.

This is a vast improvement for filter-first performance: going from
minutes to read through a multi-gigabyte file to several seconds.

This change also fixes the shutdown of the file reading tasks: we
were correctly timing out the web request but the log file reading
tasks were still persisting.

This change also increases the timeout for each file reading approach

- from 5 seconds to 10 seconds for tail-first
- from 5 seconds to 600 seconds for filter-first

Fixes #1