Show a duration timer while busy reading logs
One of my biggest pet peeves with web UX is when the system is working
and doesn't indicate to the user how long it has been working.

A duration timer isn't difficult to implement and so much nicer for
the user to be reference.

Even better is when the system itself monitors the timer and shows
relevant information or even automatically cancels requests that are
clearly out of bounds.

But in this commit we start with simply a duration timer.